Some fixes
- Add Midnight support
- Use protobuf-net for waypoints serialization
- Some QOL fixes for waypoints creator

+++ WoWDeveloperAssistant/DB2/Structures/AreaTable.cs
@@ -19,7 +19,7 @@
         public byte FactionGroupMask;
         public float AmbientMultiplier;
         public int MountFlags;
-        public short PvpCombatWorldStateId;
+        public int PvpCombatWorldStateId;
         public byte WildBattlePetLevelMin;
         public byte WildBattlePetLevelMax;
         public byte WindSettingsId;
