+++ WoWDeveloperAssistant/DB2/Structures/QuestV2.cs
@@ -3,7 +3,7 @@
     [Hotfix("QuestV2")]
     public sealed class QuestV2
     {
-        public ushort UniqueBitFlag;
+        public int UniqueBitFlag;
         public int UiQuestDetailsThemeId;
     }
 }
