+++ WoWDeveloperAssistant/DB2/Structures/SpellRange.cs
@@ -5,7 +5,7 @@
     {
         public string DisplayName;
         public string DisplayNameShort;
-        public byte Flags;
+        public int Flags;
         public float[] RangeMin = new float[2];
         public float[] RangeMax = new float[2];
     }
